+++ src/main/kotlin/de/darkatra/bfme2/ini/declarations/SageEngineIniDeclarationIndex.kt
@@ -96,7 +96,7 @@ class SageEngineIniDeclarationIndex : FileBasedIndexExtension<String, Declaratio
 
     override fun getValueExternalizer(): DataExternalizer<DeclarationOccurrence> = OccurrenceExternalizer
 
-    override fun getVersion(): Int = 6
+    override fun getVersion(): Int = 7
 
     override fun getInputFilter(): FileBasedIndex.InputFilter = object : DefaultFileTypeSpecificInputFilter(SageEngineIniFileType) {